% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polcanR.R
\name{polcanr_data}
\alias{polcanr_data}
\title{polcanr_data}
\usage{
polcanr_data(dates = polcan_date, languages = polcan_language)
}
\arguments{
\item{dates}{Dates from 2021-08-15 to 2021-09-20}

\item{languages}{Language of tweets "fr" = french, "en" = english}
}
\value{
Tweets for the dates and languages requested
}
\description{
This function allows you to retrieve tweets of the canadian election 2021 according to the selected parameters.
If no arguments are filled, all tweets will be displayed.
}
\examples{
all_tweets <- polcanr_data()
en_tweets <- polcanr_data(languages = "en")
fr_tweets <- polcanr_data(languages = "fr")
tweets_19sept <- polcanr_data(dates = "2021-09-19")
en_tweets_19sept <- polcanr_data(dates = "2021-09-19", languages = "en")
}
